Move gardener internal util to kubernetes internal
